# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/patients/otp/send": {
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "send patient otp to phone number",
                "description": "send patient otp via to phone number via sms example +213684759496",
                "operationId": "sendOtp",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+213664419425"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The otp sended successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/patients/otp/verify": {
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "verify patient otp code if match to login",
                "description": "verify patient otp code if match to login using the phone_number and the otp",
                "operationId": "loginWithOtp",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+213664419425"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "55555"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The verification passed successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Your OTP Or Phone Number is not correct",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "419": {
                        "description": "Your OTP has been expired",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/patients/complete-medical-record": {
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "update or complete patient medical record",
                "description": "update or complete patient medical record ",
                "operationId": "storePatientData",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "ahmed amine"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "25-05-1995"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 28
                                    },
                                    "gender": {
                                        "type": "integer",
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "integer",
                                        "example": "adrar alg"
                                    },
                                    "height": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "55.5"
                                    },
                                    "allergy_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "chronic_diseases_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "family_history_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "has_cancer_screening": {
                                        "type": "boolean",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "has_depression_screening": {
                                        "type": "boolean",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "other_problems": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient Record Completed Successfuly",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found with the given token or phone number, please login again",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/patients/{id}/update-phone-number": {
            "put": {
                "tags": [
                    "patients"
                ],
                "summary": "update patient phone number",
                "description": "update patient phone_number ",
                "operationId": "updatePhone",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Patient id",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+213648952765"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phone Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No Patient Found with the given phone number",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/patients/{id}": {
            "delete": {
                "tags": [
                    "patients"
                ],
                "summary": "delete patient account",
                "description": "delete patient account",
                "operationId": "deletePatientAccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Patient id to delete",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You account was deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No Patient Found with the given phone number",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}